Remove discord.commands.errors
